+++ backend/facebook-integration/src/main/java/com/example/facebook_integration/model/Topic.java
@@ -7,17 +7,19 @@ import jakarta.persistence.Id;
 
 @Entity
 public class Topic {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
+
     private String topic;
 
     // Getters and setters
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
