+++ backend/facebook-integration/src/main/java/com/example/facebook_integration/repository/TopicRepository.java
@@ -3,6 +3,6 @@ package com.example.facebook_integration.repository;
 import com.example.facebook_integration.model.Topic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TopicRepository extends JpaRepository<Topic, Long> {
+public interface TopicRepository extends JpaRepository<Topic, Integer> {
     Topic findTopByOrderByIdDesc();
 }
